# modules/patriot_gpt/data/openapi.yaml
openapi: 3.1.0
info:
  title: Patriot Odoo GPT Connector
  description: Generic API to interact with Odoo v19 Models via ChatGPT.
  version: 1.0.0
servers:
  - url: https://patriotsigns.odoo.com
    description: Production Odoo Server

components:
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic

  schemas:
    Record:
      type: object
      description: A generic Odoo record dictionary.
      properties:
        id:
          type: integer
          description: Database ID of the record
        display_name:
          type: string
          description: Display name of the record
      additionalProperties: true
      example: {"id": 1, "display_name": "Example Record", "other_field": "value"}
    
    RecordList:
      type: array
      items:
        $ref: '#/components/schemas/Record'

    CreatePayload:
      type: object
      description: Fields to set on the record.
      properties:
        name:
          type: string
          description: Name or description of the record (generic fallback)
      additionalProperties: true
      example: {"name": "New Record", "active": true}

    CreateResult:
      type: object
      properties:
        id:
          type: integer
        display_name:
          type: string
        result:
          type: string

security:
  - BasicAuth: []

paths:
  /api/gpt/{model}:
    get:
      summary: Search Records
      operationId: searchRecords
      parameters:
        - name: model
          in: path
          required: true
          schema:
            type: string
          description: Technical name of the model (e.g., res.partner)
        - name: domain
          in: query
          schema:
            type: string
            default: "[]"
          description: Odoo domain as a JSON string list
        - name: fields
          in: query
          schema:
            type: string
            default: "[]"
          description: List of fields to return
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
          description: Max number of records to return
        - name: order
          in: query
          schema:
            type: string
          description: Sort order
      responses:
        '200':
          description: List of records
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecordList'
        '401':
          description: Unauthorized
        '404':
          description: Model not found

    post:
      summary: Create Record
      operationId: createRecord
      parameters:
        - name: model
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePayload'
      responses:
        '200':
          description: Record created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateResult'

  /api/gpt/{model}/{id}:
    get:
      summary: Read One Record
      operationId: readRecord
      parameters:
        - name: model
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: fields
          in: query
          schema:
            type: string
            default: "[]"
      responses:
        '200':
          description: Record data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Record'

    put:
      summary: Update Record
      operationId: updateRecord
      parameters:
        - name: model
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePayload'
      responses:
        '200':
          description: Update successful
          content:
             application/json:
               schema:
                 $ref: '#/components/schemas/CreateResult'